Added New Hook for RowIntegrity and Improved FingerPrint and Trackable Entity

+++ src/DNTFrameworkCore/Application/Models/Id.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace DNTFrameworkCore.Application.Models
 {
-    public class Id : Id<long>
+    public class Id : Id<int>
     {
     }
